pyproject.toml
[project]
name = "tiny-oidc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask-sqlalchemy>=3.1.1",
    "flask>=3.1.0",
    "pyjwt>=2.10.1",
    "cryptography>=44.0.0",
]
